Moved Slug component into DatasetDetails
This makes it easier to manage transitions on the details page. For
example the loading state is now using Skeleton components so that the
page looks like it is loading as opposed to showing a 404 until the
content loads. It also makes it easier to remodel the contents of that
page without worrying about the Home page component.

Also fixed the link between the dataset title on the home page and the
dataset details page. It was reloading the site completely instead of
just rendering the new page. This forced a reload of the datasets list
and slowed everything down.