Fixes Headers in analysis results data table.
The cypher query needed to be reversed when the order of the organelles
selected was reversed, otherwise the headers would not match the data in
the columns returned and the table would show the incorrect results.
This does have the side effect of placing the organelle columns in the
table in the wrong order, if they are selected in reverse from the
connections stored in the database, but I don't think that will be too
upsetting for anyone.