Dataset filter now user OR for items in a category
Rather than have all the items that are selected required for the
dataset to be present in the filtered list, the new filtering system
checks for at least one item in each category. If at least one item in
each of the selected categories is not present, then the dataset will be
filtered out of the list.