Fix: remarkGfm parsing of Markdown was not working
The remarkGfm plugin was being used in the NewsPost.tsx component, but
this was too late as it was being applied after the 'remarkEmbedder' and
'html' plugins were used in the posts api. This meant that the markdown
had already been modified into html and was no longer suitable for
parsing with the remarkGfm plugin. I added the remarkGfm plugin to the
post api, so no it is applied before the html plugin.